# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpdesk API",
        "description": "This API provides features to streamline the management of support requests and user information in a helpdesk system.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "List all tickets",
                "operationId": "0082053c3590628e674a11ec0d1466e6",
                "responses": {
                    "200": {
                        "description": "Tickets list"
                    },
                    "404": {
                        "description": "No tickets found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update a specific ticket",
                "operationId": "002df222d9db8e20ad342465fc3fac03",
                "requestBody": {
                    "description": "Ticket data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Ticket ID",
                                        "type": "string"
                                    },
                                    "user_requester_id": {
                                        "description": "Requester user ID",
                                        "type": "string"
                                    },
                                    "user_assigned_id": {
                                        "description": "Assigned user ID",
                                        "type": "string"
                                    },
                                    "category": {
                                        "description": "Category",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "in progress",
                                            "resolved"
                                        ]
                                    },
                                    "priority": {
                                        "description": "Priority",
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket updated successfully"
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket",
                "operationId": "93882881cca9046d8c6ddbbab7309b4b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_requester_id": {
                                        "type": "string"
                                    },
                                    "user_assigned_id": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Show a specific ticket",
                "operationId": "62091a82e5103c19191c4c1ebd30ec6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "ticket": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete a specific ticket",
                "operationId": "e9175177b0b30a20f066103cc00aa2a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket deleted"
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": " List all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "201": {
                        "description": "Users list"
                    },
                    "404": {
                        "description": "No users found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": [
                                            "requester",
                                            "assigned"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}